Use size_t in the strings replace helper (#3680)

+++ utils.c
@@ -249,7 +249,7 @@ char *janus_string_replace(char *message, const char *old_string, const char *ne
 			tmp = strstr(pos, old_string);
 			pos = tmp;
 		}
-		uint16_t old_stringlen = strlen(outgoing)+1, new_stringlen = old_stringlen + diff*counter;
+		size_t old_stringlen = strlen(outgoing)+1, new_stringlen = old_stringlen + diff*counter;
 		if(diff > 0) {	/* Resize now */
 			tmp = g_realloc(outgoing, new_stringlen);
 			outgoing = tmp;
@@ -258,13 +258,13 @@ char *janus_string_replace(char *message, const char *old_string, const char *ne
 		pos = strstr(outgoing, old_string);
 		while(pos) {
 			if(diff > 0) {	/* Move to the right (new_string is larger than old_string) */
-				uint16_t len = strlen(pos)+1;
+				size_t len = strlen(pos)+1;
 				memmove(pos + diff, pos, len);
 				memcpy(pos, new_string, strlen(new_string));
 				pos += strlen(new_string);
 				tmp = strstr(pos, old_string);
 			} else {	/* Move to the left (new_string is smaller than old_string) */
-				uint16_t len = strlen(pos - diff)+1;
+				size_t len = strlen(pos - diff)+1;
 				memmove(pos, pos - diff, len);
 				memcpy(pos, new_string, strlen(new_string));
 				pos += strlen(old_string);
